Avances en el paquete modelo
